Hide "not ready" message after fade out
When the action of fading out is done on the actor "Not ready", the actor is hidden.
The hide() method set the bounds to x=0, y=0, width=0, height=0, so the actor is not just faded out, it is moved out of the window.
This avoid that the invisible message intercepts some user clicks.

Issue #60