Remove logic in screen
The logic code in the screen was moved to the fight logic class.
The screen convert the coordinates when a click event is being processed by an input processor.
It determines if it's a character or an icon which was clicked (using the different stages of the screen).
Then it delegates the logic code to the fight logic class.

Issue #60