Save/load state of the game
The game state is saved when a fight is triggered or when the game is closed from overworld.
For now, only position of the character and the current screen are saved (but it's always overworld screen at the moment)
The "Continue" button is displayed only if a game state is actually saved.
The "Start" button erase the current game state.